+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## vNext
 
+### Apollo Client (vNext)
+
+- Apollo Client no longer deep freezes query results.
+  [@hwillson](https://github.com/hwillson) in [#](https://github.com/apollographql/apollo-client/pull/)
+
 ### Apollo Utilities (vNext)
 
 - Replace the custom `cloneDeep` implementation with